+++ IntegrationSmoke/Package.swift
@@ -8,7 +8,7 @@ let package = Package(
         .library(name: "PaydirtIntegrationSmoke", targets: ["PaydirtIntegrationSmoke"]),
     ],
     dependencies: [
-        .package(path: ".."),
+        .package(name: "paydirt-sdk", path: ".."),
         .package(url: "https://github.com/RevenueCat/purchases-ios.git", exact: "5.81.3"),
         .package(url: "https://github.com/superwall/Superwall-iOS.git", exact: "4.16.1"),
     ],
@@ -16,7 +16,7 @@ let package = Package(
         .target(
             name: "PaydirtIntegrationSmoke",
             dependencies: [
-                .product(name: "Paydirt", package: "ios-sdk"),
+                .product(name: "Paydirt", package: "paydirt-sdk"),
                 .product(name: "RevenueCat", package: "purchases-ios"),
                 .product(name: "SuperwallKit", package: "Superwall-iOS"),
             ]